# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  description: |-
    This document describes the API is for the 
    WASAText application.
  version: 1.0.2
paths:
...
  /user:
    get:
      operationId: getUser
      summary: get all the user's details
      description: |-
        Return the authenticated user's details.
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: "Unauthorized"
        "404":
          description: "User not found"
        "500":
          description: "Internal server error. Please try again later."
...
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      summary: Get conversation details
      description: |
        Retrieves metadata for a specific conversation, including its type, name, and photo.
        If the conversation is a group, the list of participants will also be included.
      parameters:
        - $ref: "#/components/parameters/conversationId"
      responses:
        "200":
          description: Successfully retrieved conversation details
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: integer
                    description: Unique conversation identifier
                    example: 2
                  conversation_type:
                    type: string
                    description: Type of the conversation
                    enum: ["group", "private"]
                    example: "group"
                  display_name:
                    type: string
                    description: Name of the group or the other participant (for private chats)
                    example: "WASAChat"
                  display_photo_url:
                    type: string
                    description: URL of the conversation photo
                    example: "/service/photos/groups/group_2.jpg"
                  participants:
                    type: array
                    description: List of participants (only for group conversations)
                    items:
                      $ref: "#/components/schemas/User"
        "400":
          description: Invalid conversation ID
        "404":
          description: Conversation not found
        "500":
          description: Internal server error
...
  /conversations/{conversationId}/members:
    post:
      operationId: addToGroup
      summary: Add users to a group conversation
      description: |
        Adds new participants to an existing group conversation.
        The request must be sent by an existing member of the group.
      tags:
        - Conversations
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participants:
                  type: array
                  items:
                    type: integer
                    description: ID of the user to be added
                  example: [3, 4, 5]
      responses:
        "200":
          description: Successfully added users to the group conversation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: integer
                    description: Unique identifier of the conversation
                    example: 1
                  added_users:
                    type: array
                    items:
                      type: integer
                      description: ID of users that were added
                    example: [3, 4, 5]
        "400":
          description: |
            - Invalid conversation ID
            - No participants provided
            - Conversation is not a group
        "403":
          description: User is not a member of the conversation
        "404":
          description: Conversation not found
        "500":
          description: Internal server error
...
  /conversations/{conversationId}/messages:
    post:
      operationId: sendMessage
      summary: Send a message in a conversation
      description: |
        Sends a message (either text or photo) in the specified conversation.
        Users can send a plain text message, a photo, or both.
      tags:
        - Messages
      parameters:
        - $ref: "#/components/parameters/conversationId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: The text content of the message
                  example: "Hello everyone!"
                photo:
                  type: string
                  format: binary
                  description: The image file to be uploaded
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message_id:
                    type: integer
                    description: Unique identifier of the sent message
                    example: 123
                  message_type:
                    type: string
                    enum: ["text", "photo"]
                    description: Type of message sent
                    example: "text"
                  content:
                    type: string
                    description: The text content of the message (empty if it's a photo message)
                    example: "Hello everyone!"
                  sender_id:
                    type: integer
                    description: ID of the user who sent the message
                    example: 2
                  sender_name:
                    type: string
                    description: Name of the user who sent the message
                    example: "JohnDoe"
                  timestamp:
                    type: string
                    format: date-time
                    description: Time when the message was sent
                    example: "2024-02-02T15:04:05Z"
        "400":
          description: Bad request - Either no content provided, invalid file type, or malformed request.
        "401":
          description: Unauthorized - User is not authenticated.
        "404":
          description: Conversation not found.
        "500":
          description: Internal server error.
components:
  schemas:
    User:
      title: User
      description: "This object represents a single User"
      type: object
      properties:
        id:
          description: "Unique user identifier"
          type: integer
          example: 1
        username:
          type: string
          description: "The user's username"
          example: Maria
          pattern: "^[a-zA-Z0-9]*$"
          minLength: 3
          maxLength: 16
        photo_url:
          description: "URL of user's profile picture"
          type: string
          example: "https://example.com/photo.jpg"
...
  parameters:
...
    conversationId:
      schema:
        type: integer
        example: 1
      name: conversationId
      in: path
      required: true
      description: "Conversation Id"
...